requirements.txt
tensorflow==2.15.0
tensorflow-intel==2.15.0
keras==2.15.0
numpy==1.26.4
ml-dtypes==0.2.0
mediapipe==0.10.21
opencv-python==4.9.0.80
pillow==10.3.0
streamlit==1.22.0
Pillow==9.5.0
